"Implement slot symbols & enhance UI for slot machine game"
In this commit, we introduced slot symbols (BAR, BELL, CHERRY, DIAMOND, LEMON, ORANGE, SEVEN) to increase the immersion and feel of the game. Changes in SlotLine.ts and Wheel.ts were made to incorporate the use of these symbols.

A new method ‘getValues’ is added in Wheel.ts to get the actual value of slots.

UI enhancements are made in index.vue for better alignment and appearance of the slot machine.

SlotEntry.vue now contains templates to display these symbols as images instead of mere strings.

Finally, additional slots are added in SlotWheel.vue to increase complexity and challenge.

The motive behind these changes is to make game more visually appealing and challenging.