Add highlight for winning sequence and implementation for spins
Added function getIndexFromHeight to utils/Wheel.ts and utils/SlotLine.ts to get index based on the height of SlotLine. This is used during the calculation of the winning sequence. In components/SlotEntry.vue, added a class 'highlight' that alters the background color. This is used to highlight the winning sequence during the game play. In components/SlotWheel.vue, implemented a spin function to control the spinning of the slot machine and a sleep function to pause execution for a given time. These facilitate the smooth game play and user interaction.